+++ setup/setup-arch.sh
@@ -35,11 +35,11 @@ packages=(
     "network-manager-applet"
     "imagemagick"
     "polkit-gnome"
-    "hyprshade"
     "grimblast-git"
     "pacman-contrib"
     "loupe"
     "power-profiles-daemon"
+    "hyprsunset"
     # Apps
     "waypaper"
     "swaync"