+++ setup/setup-fedora.sh
@@ -34,6 +34,7 @@ packages=(
     "ImageMagick"
     "NetworkManager-tui"
     "tesseract-langpack-eng"
+    "hyprsunset"
     # Apps
     "waypaper"
     "swww"
@@ -166,7 +167,6 @@ sudo cp $SCRIPT_DIR/packages/eza /usr/bin
 # --------------------------------------------------------------
 
 echo ":: Installing packages with pip"
-sudo pip install hyprshade
 sudo pip install pywalfox
 sudo pip install screeninfo
 sudo pip install waypaper